requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
pandas==2.2.2
numpy==1.26.4
prophet==1.1.5
scikit-learn==1.5.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
apscheduler==3.10.4
httpx==0.27.0
pydantic-settings==2.3.1
